Added to and improved lexing, added a basic AST and parser, very simple interpretter
Most of the language is lexed, some of it can be parsed now. Still need semantic analysis to make sense of types and names.

For examples, try:
stack build --force-dirty && stack exec plt < example/parsing-hello-world-example.wl

The 'interpretter' just looks for a function called main, traverses it looking for "log" calls, and prints those to stdout.